+++ .vscode/launch.json
@@ -10,9 +10,9 @@
       "request": "launch",
       "preLaunchTask": "build",
       // If you have changed target frameworks, make sure to update the program path.
-      "program": "${workspaceFolder}/services/form-metadata/test/FormMetadata.Functions.Tests/bin/Debug/net6.0/FormMetadata.Functions.Tests.dll",
+      "program": "${workspaceFolder}/services/form/test/Form.Functions.Tests/bin/Debug/net6.0/Form.Functions.Tests.dll",
       "args": [],
-      "cwd": "${workspaceFolder}/services/form-metadata/test/FormMetadata.Functions.Tests",
+      "cwd": "${workspaceFolder}/services/form/test/Form.Functions.Tests",
       // For more information about the 'console' field, see https://aka.ms/VSCode-CS-LaunchJson-Console
       "console": "internalConsole",
       "stopAtEntry": false
