--- services/form-metadata/FormMetadata.sln
@@ -1,57 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30114.105
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{337C271D-E62A-4BA5-9BFA-502101F2DFE7}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FormMetadata.Functions", "src\FormMetadata.Functions\FormMetadata.Functions.csproj", "{F7FE077F-1EEC-4637-B139-1B4EB12DC038}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{971FFB1A-35DE-47D5-9001-2CA4A8A18360}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FormMetadata.Functions.Tests", "test\FormMetadata.Functions.Tests\FormMetadata.Functions.Tests.csproj", "{962D614F-B381-4703-AF1E-6501385D0F29}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FormMetadata.Domain", "src\FormMetadata.Domain\FormMetadata.Domain.csproj", "{E0482CF1-330D-4D1A-AC1F-E06DE7300227}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FormMetadata.Application", "src\FormMetadata.Application\FormMetadata.Application.csproj", "{C135EC82-151C-4A47-BC8B-81356FC72A2F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FormMetadata.Infrastructure", "src\FormMetadata.Infrastructure\FormMetadata.Infrastructure.csproj", "{C60CE6E9-25EE-415A-8524-5AEC552A01B1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F7FE077F-1EEC-4637-B139-1B4EB12DC038}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F7FE077F-1EEC-4637-B139-1B4EB12DC038}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F7FE077F-1EEC-4637-B139-1B4EB12DC038}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F7FE077F-1EEC-4637-B139-1B4EB12DC038}.Release|Any CPU.Build.0 = Release|Any CPU
-		{962D614F-B381-4703-AF1E-6501385D0F29}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{962D614F-B381-4703-AF1E-6501385D0F29}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{962D614F-B381-4703-AF1E-6501385D0F29}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{962D614F-B381-4703-AF1E-6501385D0F29}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E0482CF1-330D-4D1A-AC1F-E06DE7300227}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E0482CF1-330D-4D1A-AC1F-E06DE7300227}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E0482CF1-330D-4D1A-AC1F-E06DE7300227}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E0482CF1-330D-4D1A-AC1F-E06DE7300227}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C135EC82-151C-4A47-BC8B-81356FC72A2F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C135EC82-151C-4A47-BC8B-81356FC72A2F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C135EC82-151C-4A47-BC8B-81356FC72A2F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C135EC82-151C-4A47-BC8B-81356FC72A2F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C60CE6E9-25EE-415A-8524-5AEC552A01B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C60CE6E9-25EE-415A-8524-5AEC552A01B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C60CE6E9-25EE-415A-8524-5AEC552A01B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C60CE6E9-25EE-415A-8524-5AEC552A01B1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{F7FE077F-1EEC-4637-B139-1B4EB12DC038} = {337C271D-E62A-4BA5-9BFA-502101F2DFE7}
-		{962D614F-B381-4703-AF1E-6501385D0F29} = {971FFB1A-35DE-47D5-9001-2CA4A8A18360}
-		{E0482CF1-330D-4D1A-AC1F-E06DE7300227} = {337C271D-E62A-4BA5-9BFA-502101F2DFE7}
-		{C135EC82-151C-4A47-BC8B-81356FC72A2F} = {337C271D-E62A-4BA5-9BFA-502101F2DFE7}
-		{C60CE6E9-25EE-415A-8524-5AEC552A01B1} = {337C271D-E62A-4BA5-9BFA-502101F2DFE7}
-	EndGlobalSection
-EndGlobal
