--- services/form-metadata/src/FormMetadata.Application/Queries/IQuery.cs
@@ -1,11 +1,0 @@
-namespace FormMetadata.Application.Queries;
-
-public interface IQuery
-{
-}
-
-public interface IQueryHandler<in TQuery, TResult>
-    where TQuery : IQuery
-{
-    public Task<TResult?> Handle(TQuery query);
-}
