--- services/form-metadata/src/FormMetadata.Functions/samconfig.toml
@@ -1,12 +1,0 @@
-version = 0.1
-[default]
-[default.deploy]
-[default.deploy.parameters]
-stack_name = "formchata-form-metadata"
-s3_bucket = "aws-sam-cli-managed-default-samclisourcebucket-ox5lpxf7jfcf"
-s3_prefix = "formchata-form-metadata"
-region = "us-east-1"
-confirm_changeset = true
-capabilities = "CAPABILITY_IAM"
-disable_rollback = true
-image_repositories = []
