+++ services/form/src/Form.Application/Queries/IQuery.cs
@@ -1,0 +1,11 @@
+namespace Form.Application.Queries;
+
+public interface IQuery
+{
+}
+
+public interface IQueryHandler<in TQuery, TResult>
+    where TQuery : IQuery
+{
+    public Task<TResult?> Handle(TQuery query);
+}
