+++ services/form/test/Form.Functions.Tests/CreateFormTests.cs
@@ -1,0 +1,16 @@
+using Amazon.Lambda.APIGatewayEvents;
+using Amazon.Lambda.TestUtilities;
+using Xunit;
+
+namespace Form.Functions.Tests;
+
+public class CreateFormTests
+{
+    public CreateFormTests()
+    {
+    }
+
+    public void TestHandle()
+    {
+    }
+}
